Moved spot and profile routes into their own files
